refactor distance functions into separate module
